+++ components/apps/src/lib.rs
@@ -66,6 +66,9 @@ pub struct Config {
     fido: FidoConfig,
     #[serde(default, rename = "o", skip_serializing_if = "is_default")]
     opcard: OpcardConfig,
+    #[cfg(feature = "piv-authenticator")]
+    #[serde(default, rename = "p", skip_serializing_if = "is_default")]
+    piv: PivConfig,
     #[serde(default, rename = "v", skip_serializing_if = "is_default")]
     fs_version: u32,
     #[cfg(feature = "se050")]
@@ -79,6 +82,8 @@ impl admin_app::Config for Config {
         match app {
             "fido" => self.fido.field(key),
             "opcard" => self.opcard.field(key),
+            #[cfg(feature = "piv-authenticator")]
+            "piv" => self.piv.field(key),
             _ => None,
         }
     }
@@ -100,6 +105,23 @@ impl admin_app::Config for Config {
                 destructive: true,
                 ty: FieldType::Bool,
             },
+            ConfigField {
+                name: "opcard.disabled",
+                requires_touch_confirmation: false,
+                // APDU dispatch does not handle well having the currently select application removed
+                requires_reboot: true,
+                destructive: false,
+                ty: FieldType::Bool,
+            },
+            #[cfg(feature = "piv-authenticator")]
+            ConfigField {
+                name: "piv.disabled",
+                requires_touch_confirmation: false,
+                // APDU dispatch does not handle well having the currently select application removed
+                requires_reboot: true,
+                destructive: false,
+                ty: FieldType::Bool,
+            },
         ]
     }
 
@@ -114,6 +136,11 @@ impl admin_app::Config for Config {
 
             (Some(("opcard", key)), _) => self.opcard.reset_client_id(key),
             (None, "opcard") => self.opcard.reset_client_id(""),
+
+            #[cfg(feature = "piv-authenticator")]
+            (Some(("piv", key)), _) => self.piv.reset_client_id(key),
+            #[cfg(feature = "piv-authenticator")]
+            (None, "piv") => self.piv.reset_client_id(""),
 
             _ => None,
         };
@@ -185,6 +212,8 @@ pub struct OpcardConfig {
     #[cfg(feature = "se050")]
     #[serde(default, rename = "s", skip_serializing_if = "is_default")]
     use_se050_backend: bool,
+    #[serde(default, rename = "d", skip_serializing_if = "is_default")]
+    disabled: bool,
 }
 
 #[cfg(feature = "opcard")]
@@ -222,6 +251,7 @@ impl OpcardConfig {
         Self {
             #[cfg(feature = "se050")]
             use_se050_backend: true,
+            disabled: false,
         }
     }
 
@@ -229,6 +259,7 @@ impl OpcardConfig {
         match key {
             #[cfg(feature = "se050")]
             "use_se050_backend" => Some(ConfigValueMut::Bool(&mut self.use_se050_backend)),
+            "disabled" => Some(ConfigValueMut::Bool(&mut self.disabled)),
             _ => None,
         }
     }
@@ -258,6 +289,31 @@ impl OpcardConfig {
             ResetConfigResult::Changed
         }
     }
+}
+
+#[cfg(feature = "piv-authenticator")]
+impl PivConfig {
+    fn field(&mut self, key: &str) -> Option<ConfigValueMut<'_>> {
+        match key {
+            "disabled" => Some(ConfigValueMut::Bool(&mut self.disabled)),
+            _ => None,
+        }
+    }
+
+    #[cfg(feature = "factory-reset")]
+    fn reset_client_id(
+        &self,
+        _key: &str,
+    ) -> Option<(&'static Path, &'static ResetSignalAllocation)> {
+        None
+    }
+}
+
+#[cfg(feature = "piv-authenticator")]
+#[derive(Debug, PartialEq, Deserialize, Serialize, Default)]
+pub struct PivConfig {
+    #[serde(default, rename = "d", skip_serializing_if = "is_default")]
+    disabled: bool,
 }
 
 pub trait Runner {
@@ -599,12 +655,16 @@ impl<R: Runner> Apps<R> {
 
         #[cfg(feature = "opcard")]
         if let Some(opcard) = self.opcard.as_mut() {
-            apps.push(opcard).ok().unwrap();
+            if !self.admin.config().opcard.disabled {
+                apps.push(opcard).ok().unwrap();
+            }
         }
 
         #[cfg(feature = "piv-authenticator")]
         if let Some(piv) = self.piv.as_mut() {
-            apps.push(piv).ok().unwrap();
+            if !self.admin.config().piv.disabled {
+                apps.push(piv).ok().unwrap();
+            }
         }
 
         #[cfg(all(feature = "fido-authenticator", not(feature = "webcrypt")))]
@@ -1124,6 +1184,8 @@ impl<R: Runner> App<R> for ProvisionerApp<R> {
 
 #[cfg(test)]
 mod tests {
+    #[cfg(feature = "piv-authenticator")]
+    use super::PivConfig;
     use super::{Config, FidoConfig, OpcardConfig};
     use cbor_smol::cbor_serialize_bytes;
 
@@ -1136,7 +1198,10 @@ mod tests {
             opcard: OpcardConfig {
                 #[cfg(feature = "se050")]
                 use_se050_backend: true,
+                disabled: true,
             },
+            #[cfg(feature = "piv-authenticator")]
+            piv: PivConfig { disabled: true },
             fs_version: 1,
         };
         let data: heapless_bytes::Bytes<1024> = cbor_serialize_bytes(&config).unwrap();